8-deploy-langserve/requirements.txt
langchain
langchain-openai
langchain-community
langchain-text-splitters
langchain-postgres
python-dotenv
beautifulsoup4
requests
pypdf
psycopg
psycopg[binary]
langgraph
ddgs
wikipedia
fastapi
uvicorn
numexpr
sse_starlette
langserve
